+++ streamable/_afunctions.py
@@ -3,7 +3,6 @@ import datetime
 from contextlib import suppress
 from inspect import iscoroutinefunction
 from typing import (
-    TYPE_CHECKING,
     Any,
     AsyncIterable,
     AsyncIterator,
@@ -17,9 +16,6 @@ from typing import (
     Union,
     cast,
 )
-
-if TYPE_CHECKING:
-    from streamable._stream import stream
 
 from streamable import _aiterators
 from streamable._tools._async import AsyncCallable
@@ -123,20 +119,15 @@ def observe(
     aiterator: AsyncIterator[T],
     subject: str,
     every: Union[None, int, datetime.timedelta],
-    do: Union[
-        None,
-        Callable[["stream.Observation"], Any],
-        AsyncCallable["stream.Observation", Any],
-    ] = None,
 ) -> AsyncIterator[T]:
     if every is None:
-        return _aiterators.PowerObserveAsyncIterator(aiterator, subject, asyncify(do))
+        return _aiterators.PowerObserveAsyncIterator(aiterator, subject, do=None)
     elif isinstance(every, int):
         return _aiterators.EveryIntObserveAsyncIterator(
-            aiterator, subject, every, asyncify(do)
+            aiterator, subject, every, do=None
         )
     return _aiterators.EveryIntervalObserveAsyncIterator(
-        aiterator, subject, every, asyncify(do)
+        aiterator, subject, every, do=None
     )