+++ streamable/_aiterators.py
@@ -22,6 +22,7 @@ from typing import (
     Iterable,
     Iterator,
     List,
+    NamedTuple,
     Optional,
     Tuple,
     Type,
@@ -31,7 +32,7 @@ from typing import (
 )
 
 if TYPE_CHECKING:
-    from streamable._stream import stream
+    pass
 from streamable._tools._afuture import (
     FutureResult,
     FutureResultCollection,
@@ -372,11 +373,17 @@ class FilterAsyncIterator(AsyncIterator[T]):
 class ObserveAsyncIterator(AsyncIterator[T]):
     _FORMAT = "stream={subject} elapsed={elapsed} errors={errors} emissions={emissions}"
 
+    class Observation(NamedTuple):
+        subject: str
+        elapsed: datetime.timedelta
+        errors: int
+        emissions: int
+
     def __init__(
         self,
         iterator: AsyncIterator[T],
         subject: str,
-        do: Optional[AsyncCallable["stream.Observation", Any]],
+        do: Optional[AsyncCallable[Observation, Any]],
     ) -> None:
         self.iterator = iterator
         self.subject = logfmt_str_escape(subject)
@@ -388,13 +395,11 @@ class ObserveAsyncIterator(AsyncIterator[T]):
         self._errors_logged = 0
         self.__start_point: Optional[datetime.datetime] = None
 
-    async def _log(self, observation: "stream.Observation") -> None:
+    async def _log(self, observation: Observation) -> None:
         get_logger().info(self._FORMAT.format(**observation._asdict()))
 
-    def _observation(self) -> "stream.Observation":
-        from streamable._stream import stream
-
-        return stream.Observation(
+    def _observation(self) -> Observation:
+        return self.Observation(
             subject=self.subject,
             elapsed=self._time_point() - self._start_point(),
             errors=self._errors,
@@ -446,7 +451,7 @@ class PowerObserveAsyncIterator(ObserveAsyncIterator[T]):
         self,
         iterator: AsyncIterator[T],
         subject: str,
-        do: Optional[AsyncCallable["stream.Observation", Any]],
+        do: Optional[AsyncCallable[ObserveAsyncIterator.Observation, Any]],
         base: int = 2,
     ) -> None:
         super().__init__(iterator, subject, do)
@@ -465,7 +470,7 @@ class EveryIntObserveAsyncIterator(ObserveAsyncIterator[T]):
         iterator: AsyncIterator[T],
         subject: str,
         every: int,
-        do: Optional[AsyncCallable["stream.Observation", Any]],
+        do: Optional[AsyncCallable[ObserveAsyncIterator.Observation, Any]],
     ) -> None:
         super().__init__(iterator, subject, do)
         self.every = every
@@ -485,7 +490,7 @@ class EveryIntervalObserveAsyncIterator(ObserveAsyncIterator[T]):
         iterator: AsyncIterator[T],
         subject: str,
         every: datetime.timedelta,
-        do: Optional[AsyncCallable["stream.Observation", Any]],
+        do: Optional[AsyncCallable[ObserveAsyncIterator.Observation, Any]],
     ) -> None:
         super().__init__(iterator, subject, do)
         self._every_seconds: float = every.total_seconds()