+++ streamable/visitors/_eq.py
@@ -84,7 +84,6 @@ class EqualityVisitor(Visitor[bool]):
             and stream.upstream.accept(EqualityVisitor(self.other.upstream))
             and stream._subject == self.other._subject
             and stream._every == self.other._every
-            and stream._do == self.other._do
         )
 
     def visit_skip_stream(self, stream: "SkipStream") -> bool: