+++ tests/operations/test_observe.py
@@ -5,18 +5,16 @@ Observe logs iteration progress including errors and emissions at specified inte
 """
 
 import datetime
-from typing import Any, Callable, Iterable, List, NamedTuple, Union
+from typing import Iterable, List, NamedTuple, Union
 from unittest.mock import patch
 
 import pytest
 
 from streamable import stream
-from streamable._tools._func import asyncify
 from streamable._tools._logging import logfmt_str_escape
 from tests.utils import (
     ITERABLE_TYPES,
     IterableType,
-    identity,
     slow_identity,
     slow_identity_duration,
     to_list,
@@ -436,27 +434,27 @@ def test_observe_logging_every_timedelta_frequent(itype: IterableType) -> None:
         assert len(digits) == len(logs)
 
 
-# ============================================================================
-# Custom How Function Tests
-# ============================================================================
+# # ============================================================================
+# # Custom How Function Tests
+# # ============================================================================
 
 
-@pytest.mark.parametrize("adapt", [identity, asyncify])
-@pytest.mark.parametrize("itype", ITERABLE_TYPES)
-def test_observe_do(
-    itype: IterableType, adapt: Callable[[Callable[[Any], Any]], Callable[[Any], Any]]
-) -> None:
-    """Observe should call the do function with log messages."""
-    observations: List[stream.Observation] = []
-    ints = list(range(8))
-    assert (
-        to_list(
-            stream(ints).observe("ints", every=2, do=adapt(observations.append)),
-            itype,
-        )
-        == ints
-    )
-    assert [observation.emissions for observation in observations] == [1, 2, 4, 6, 8]
+# @pytest.mark.parametrize("adapt", [identity, asyncify])
+# @pytest.mark.parametrize("itype", ITERABLE_TYPES)
+# def test_observe_do(
+#     itype: IterableType, adapt: Callable[[Callable[[Any], Any]], Callable[[Any], Any]]
+# ) -> None:
+#     """Observe should call the do function with log messages."""
+#     observations: List[stream._Observation] = []
+#     ints = list(range(8))
+#     assert (
+#         to_list(
+#             stream(ints).observe("ints", every=2, do=adapt(observations.append)),
+#             itype,
+#         )
+#         == ints
+#     )
+#     assert [observation.emissions for observation in observations] == [1, 2, 4, 6, 8]
 
 
 # ============================================================================